modified:   MidiCode/classdefs.cpp        modified:   MidiCode/compileclass.sh 	deleted:    MidiCode/convertgdbmidiclass 	modified:   MidiCode/convertgdbmidiclass.cpp 	modified:   MidiCode/libconvertgdbmidiclass.cpp 	modified:   MidiCode/libdefs.h 	modified:   MidiCode/libmidiclass.h 	modified:   MidiCode/midiclass.h 	new file:   MidiCode/midiout.cpp 	new file:   MidiCode/parse.cpp 	new file:   MidiCode/returnstatus.cpp 	new file:   MidiCode/utils.cpp

+++ gbell/MidiCode/compileclass.sh
@@ -2,10 +2,43 @@
 
 set -x
 
+rm *.o
 g++ -g -c convertgdbmidiclass.cpp
+if [ $? -ne 0 ]
+then
+    exit 1
+fi
 g++ -g -c libconvertgdbmidiclass.cpp
+if [ $? -ne 0 ]
+then
+    exit 1 
+fi
 g++ -g -c classdefs.cpp
-g++ -g -o convertgdbmidiclass convertgdbmidiclass.o libconvertgdbmidiclass.o classdefs.o
+if [ $? -ne 0 ]
+then
+    exit 1 
+fi
+g++ -g -c returnstatus.cpp
+if [ $? -ne 0 ]
+then
+    exit 1 
+fi
+g++ -g -c parse.cpp
+if [ $? -ne 0 ]
+then
+    exit 1 
+fi
+g++ -g -c utils.cpp
+if [ $? -ne 0 ]
+then
+    exit 1 
+fi
+g++ -g -c midiout.cpp
+if [ $? -ne o ]
+then
+    exit 1
+fi
+g++ -g -o convertgdbmidiclass convertgdbmidiclass.o libconvertgdbmidiclass.o classdefs.o returnstatus.o parse.o utils.cpp midiout.cpp
 if [ $? -eq 0 ]
 then
     rm *.o
